Robust adaptive lasso based on robust correlation estimates

Robust adaptive lasso based on robust correlation estimates

Usage

covlasso(
  x,
  y,
  cor.method = "gaussrank",
  scale.method = "qn",
  center.method = "median",
  adaptive = TRUE
)

Arguments

x

input design matrix

y

input response vector

cor.method

"sd" or "qn"

scale.method

could be "pearson", "pair" and "gaussrank"

adaptive

adaptive regularization penalties

pda.method

"nearpd" or FALSE

lmin

min threshold of eigenvalues

std

If TRUE the robust correlation matrix is used, if FALSE the robust covariance matrix is used.

cormatrix

you could also use a correlation matrix as input

scale

put in scales if you use cormatrix

Value

betahat_opt, the optimal estimation of beta obtained from this algorithm

lambda_opt is the optimal tuning parameter and sigma_opt is the optimal estimation of sigma.

The output also includes the estimated correlation matrix, the estimated covariance matrix and et cetera from the covf function.

Examples

dat = genevar()
y = dat$y
x = dat$x
fit = covlasso(x,y)
fit$betahat_opt
#>    const.3      beta1      beta2      beta3      beta4      beta5      beta6 
#> -0.5731841  0.0000000  1.5415224  1.5957187  2.3790625  0.0000000  0.0000000 
#>      beta7      beta8      beta9     beta10     beta11     beta12     beta13 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 
#>     beta14     beta15     beta16     beta17     beta18     beta19     beta20 
#>  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000