Robust adaptive lasso based on robust correlation estimates

Robust adaptive lasso based on robust correlation estimates

covlasso(
  x,
  y,
  cor.method = "pair",
  scale.method = "qn",
  pda.method = NULL,
  lmin = NULL,
  std = T,
  adaptive = T,
  cormatrix = NULL,
  scale = NULL
)

Arguments

x
input design matrix

y
input response vector

cor.method
"sd" or "qn"

scale.method
could be "pearson", "pair" and "gaussrank"

pda.method
"nearpd" or FALSE

lmin
min threshold of eigenvalues

std
If TRUE the robust correlation matrix is used, if FALSE the robust covariance matrix is used.

adaptive
adaptive regularization penalties

cormatrix
you could also use a correlation matrix as input

scale
put in scales if you use cormatrix

Value

betahat_opt, the optimal estimation of beta obtained from this algorithm

lambda_opt is the optimal tuning parameter and sigma_opt is the optimal estimation of sigma.

The output also includes the estimated correlation matrix, the estimated covariance matrix and et cetera from the covf function.

Examples

dat = genevar()
y = dat$y
x = dat$x
fit = covlasso(x,y)
fit$betahat_opt

#>  [1] 0.7175245 1.3139470 1.3813162 1.1393393 0.7682718 0.0000000 0.0000000
#>  [8] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> [15] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000